+++ src/docs/pages/component-docs/components/drawer/drawer-nav-item/examples/basic.component.ts
@@ -1,5 +1,4 @@
-import { Component, ViewChild } from '@angular/core';
-import { DrawerComponent } from '@brightlayer-ui/angular-components';
+import { Component } from '@angular/core';
 
 export const BASIC = `<blui-drawer style="width: 250px">
     <blui-drawer-body>
@@ -16,10 +15,4 @@ export const BASIC = `<blui-drawer style="width: 250px">
     selector: 'app-basic-drawer-nav-item-demo',
     template: BASIC,
 })
-export class BasicExampleComponent {
-    @ViewChild(DrawerComponent) drawer;
-
-    ngAfterViewInit(): void {
-        this.drawer.openOnHover = false;
-    }
-}
+export class BasicExampleComponent {}