+++ src/docs/pages/component-docs/components/drawer/drawer-nav-item/examples/with-icons.component.ts
@@ -1,5 +1,4 @@
-import { Component, ViewChild } from '@angular/core';
-import { DrawerComponent } from '@brightlayer-ui/angular-components';
+import { Component } from '@angular/core';
 
 export const WITH_ICONS = `<blui-drawer style="width: 250px">
     <blui-drawer-body>
@@ -19,10 +18,4 @@ export const WITH_ICONS = `<blui-drawer style="width: 250px">
     selector: 'app-nav-item-with-icons-demo',
     template: WITH_ICONS,
 })
-export class WithIconsExampleComponent {
-    @ViewChild(DrawerComponent) drawer;
-
-    ngAfterViewInit(): void {
-        this.drawer.openOnHover = false;
-    }
-}
+export class WithIconsExampleComponent {}