+++ src/docs/pages/component-docs/components/drawer/drawer-nav-item/examples/with-nested-items.component.ts
@@ -1,5 +1,4 @@
-import { Component, ViewChild } from '@angular/core';
-import { DrawerComponent } from '@brightlayer-ui/angular-components';
+import { Component } from '@angular/core';
 
 export const WITH_NESTED_ITEMS = `<blui-drawer style="width: 250px">
     <blui-drawer-body>
@@ -23,10 +22,4 @@ export const WITH_NESTED_ITEMS = `<blui-drawer style="width: 250px">
     selector: 'app-nav-item-with-nested-items-demo',
     template: WITH_NESTED_ITEMS,
 })
-export class WithNestedItemsComponent {
-    @ViewChild(DrawerComponent) drawer;
-
-    ngAfterViewInit(): void {
-        this.drawer.openOnHover = false;
-    }
-}
+export class WithNestedItemsComponent {}