+++ src/lib/core/drawer/state-listener.component.ts
@@ -1,12 +1,50 @@
-import { ChangeDetectorRef, OnDestroy, OnInit, Directive } from '@angular/core';
-import { DrawerService } from './service/drawer.service';
+import { ChangeDetectorRef, OnDestroy, OnInit, Directive, Injectable } from '@angular/core';
 import { Subscription } from 'rxjs';
+import { DrawerState } from './drawerState';
+
+const stateMap = new Map<number, DrawerState>();
+
+/** This service is used to manage the state of a Drawer component, responds to user behavior and input settings. */
+@Injectable({
+    providedIn: 'root',
+})
+export class DrawerStateManagerService {
+    private numberOfDrawerInstances = 0;
+
+    incrementNumberOfDrawerInstances(): number {
+        this.numberOfDrawerInstances = this.numberOfDrawerInstances + 1;
+        return this.numberOfDrawerInstances;
+    }
+
+    getNumberOfDrawerInstances(): number {
+        return this.numberOfDrawerInstances;
+    }
+}
 
 @Directive()
 export class StateListener implements OnInit, OnDestroy {
     drawerOpenListener: Subscription;
+    drawerState: DrawerState;
 
-    constructor(protected drawerService: DrawerService, protected changeDetector: ChangeDetectorRef) {}
+    constructor(
+        protected drawerStateManagerService: DrawerStateManagerService,
+        protected changeDetector: ChangeDetectorRef,
+        public createNewStateListener = false
+    ) {
+        if (createNewStateListener || !stateMap.has(drawerStateManagerService.getNumberOfDrawerInstances())) {
+            this._createNewDrawerServiceInstance();
+        } else {
+            this.drawerState = stateMap.get(drawerStateManagerService.getNumberOfDrawerInstances());
+        }
+    }
+
+    private _createNewDrawerServiceInstance(): void {
+        this.drawerStateManagerService.incrementNumberOfDrawerInstances();
+        const newDrawerStateInstance = new DrawerState();
+        const drawerId = this.drawerStateManagerService.getNumberOfDrawerInstances();
+        stateMap.set(drawerId, newDrawerStateInstance);
+        this.drawerState = newDrawerStateInstance;
+    }
 
     public ngOnInit(): void {
         this.listenForDrawerChanges();
@@ -17,11 +55,12 @@ export class StateListener implements OnInit, OnDestroy {
     }
 
     public isOpen(): boolean {
-        return this.drawerService.isDrawerOpen();
+        // console.log(this.drawerService.getNumberOfDrawerInstances());
+        return this.drawerState.isDrawerOpen();
     }
 
     public isOpenOnHover(): boolean {
-        return this.drawerService.isOpenOnHover();
+        return this.drawerState.isOpenOnHover();
     }
 
     public unsubscribeListeners(): void {
@@ -31,7 +70,7 @@ export class StateListener implements OnInit, OnDestroy {
     }
 
     listenForDrawerChanges(): void {
-        this.drawerOpenListener = this.drawerService.drawerOpenChanges().subscribe(() => {
+        this.drawerOpenListener = this.drawerState.drawerOpenChanges().subscribe(() => {
             this.changeDetector.detectChanges();
         });
     }